requirements/base.txt
boto3==1.9.152
botocore==1.12.152
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-cors-headers==3.0.1
django-filter==2.1.0
django-redis==4.10.0
django-storages==1.7.1
djangorestframework==3.9.4
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
docutils==0.14
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.2.post1
pika==1.0.1
Pillow==6.0.0
PyJWT==1.7.1
pymongo==3.8.0
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
s3transfer==0.2.0
six==1.12.0
sqlparse==0.3.0
unicodecsv==0.14.1
urllib3==1.24.3